Do not use same name for App namespace and class
