+++ CDP4-SAT.sln
@@ -19,9 +19,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Migration.Tests", "Modules\Migration.Tests\Migration.Tests.csproj", "{EC675A58-6D57-4E6E-A4CA-9CE38D736E0F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "App", "App\App.csproj", "{195A601C-3905-4FD6-9C37-30F359C6D753}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Syncer.Tests", "Modules\Syncer.Tests\Syncer.Tests.csproj", "{09E9476A-EA4C-4E0D-993D-5003E784C059}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syncer.Tests", "Modules\Syncer.Tests\Syncer.Tests.csproj", "{09E9476A-EA4C-4E0D-993D-5003E784C059}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SAT", "SAT\SAT.csproj", "{15228E30-401B-4FE2-9B7E-21BC9EF5D50E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -49,14 +49,14 @@ Global
 		{EC675A58-6D57-4E6E-A4CA-9CE38D736E0F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC675A58-6D57-4E6E-A4CA-9CE38D736E0F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EC675A58-6D57-4E6E-A4CA-9CE38D736E0F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{195A601C-3905-4FD6-9C37-30F359C6D753}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{195A601C-3905-4FD6-9C37-30F359C6D753}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{195A601C-3905-4FD6-9C37-30F359C6D753}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{195A601C-3905-4FD6-9C37-30F359C6D753}.Release|Any CPU.Build.0 = Release|Any CPU
 		{09E9476A-EA4C-4E0D-993D-5003E784C059}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{09E9476A-EA4C-4E0D-993D-5003E784C059}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{09E9476A-EA4C-4E0D-993D-5003E784C059}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{09E9476A-EA4C-4E0D-993D-5003E784C059}.Release|Any CPU.Build.0 = Release|Any CPU
+		{15228E30-401B-4FE2-9B7E-21BC9EF5D50E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15228E30-401B-4FE2-9B7E-21BC9EF5D50E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15228E30-401B-4FE2-9B7E-21BC9EF5D50E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15228E30-401B-4FE2-9B7E-21BC9EF5D50E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
