+++ SAT/App.xaml
@@ -1,5 +1,5 @@
 ﻿<Application
-    x:Class="App.App"
+    x:Class="SAT.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <Application.Resources>
