+++ SAT/MainWindow.xaml
@@ -1,15 +1,15 @@
 ﻿<dx:ThemedWindow
-    x:Class="App.MainWindow"
+    x:Class="SAT.MainWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:dx="http://schemas.devexpress.com/winfx/2008/xaml/core"
-    xmlns:views="clr-namespace:App.Views"
+    xmlns:views="clr-namespace:SAT.Views"
     Title="CDP4 SAT"
     Height="850"
     Width="1000"
     ResizeMode="CanMinimize"
     WindowStartupLocation="CenterScreen"
-    Icon="pack://application:,,,/App;component/Resources/cdplogo3d_48x48.png">
+    Icon="pack://application:,,,/SAT;component/Resources/cdplogo3d_48x48.png">
     <Grid>
         <views:MainView />
     </Grid>
