+++ SAT/MainWindow.xaml.cs
@@ -23,7 +23,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace App
+namespace SAT
 {
     using DevExpress.Xpf.Core;
