+++ SAT/Properties/AssemblyInfo.cs
