+++ SAT/Properties/Resources.resx
