+++ SAT/Properties/Settings.settings
